move the TestManager to the integrationBenchmark application
